requirements.txt
asgiref==3.8.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
crispy-bootstrap4==2024.10
crispy-bootstrap5==2024.10
cryptography==44.0.0
DateTime==5.5
dj-rest-auth==7.0.1
Django==4.2.18
django-allauth==65.3.1
django-browser-reload==1.17.0
django-ckeditor==6.7.2
django-crispy-forms==2.3
django-environ==0.12.0
django-filter==23.5
django-jazzmin==3.0.1
django-js-asset==3.0.1
django-phonenumber-field==8.0.0
django-resized==1.0.3
djangorestframework==3.14.0
drf-yasg==1.21.8
Faker==35.0.0
idna==3.10
inflection==0.5.1
jwt==1.3.1
Markdown==3.7
packaging==24.2
phonenumbers==8.13.53
pillow==11.1.0
pycparser==2.22
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2023.4
PyYAML==6.0.2
requests==2.32.3
setuptools==75.8.0
six==1.16.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
zope.interface==7.2
